+++ packages/admin-ui/src/lib/catalog/src/components/product-options-editor/product-options-editor.component.ts
@@ -99,6 +99,7 @@ export class ProductOptionsEditorComponent extends BaseDetailComponent<ProductWi
                 take(1),
                 mergeMap(([{ optionGroups }, languageCode, product]) => {
                     const updateOperations: Array<Observable<any>> = [];
+                    const updatedProductOptionInputs: UpdateProductOptionInput[] = [];
                     for (const optionGroupForm of this.getOptionGroups()) {
                         if (optionGroupForm.dirty) {
                             const optionGroupEntity = optionGroups.find(
@@ -127,16 +128,20 @@ export class ProductOptionsEditorComponent extends BaseDetailComponent<ProductWi
                                         optionForm,
                                         languageCode,
                                     );
-                                    updateOperations.push(
-                                        this.productDetailService.updateProductOption(
-                                            { ...input, autoUpdate: this.autoUpdateVariantNames },
-                                            product,
-                                            languageCode,
-                                        ),
-                                    );
+                                    updatedProductOptionInputs.push(input);
                                 }
                             }
                         }
+                    }
+                    if (updatedProductOptionInputs.length) {
+                        updateOperations.push(
+                            this.productDetailService.updateProductOptions(
+                                updatedProductOptionInputs,
+                                this.autoUpdateVariantNames,
+                                product,
+                                languageCode,
+                            ),
+                        );
                     }
                     return forkJoin(updateOperations);
                 }),