+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## <small>2.2.6 (2024-06-18)</small>
+
+
+#### Fixes
+
+* **admin-ui** Fix id of customer groups list ([33780ce](https://github.com/vendure-ecommerce/vendure/commit/33780ce)), closes [#2894](https://github.com/vendure-ecommerce/vendure/issues/2894)
+* **admin-ui** Fix layout issues with asset picker dialog ([996710c](https://github.com/vendure-ecommerce/vendure/commit/996710c)), closes [#2898](https://github.com/vendure-ecommerce/vendure/issues/2898)
+* **admin-ui** Fix layout of custom timeline entries ([3e075aa](https://github.com/vendure-ecommerce/vendure/commit/3e075aa))
+* **admin-ui** Fix renaming of multiple product options at once ([3de60b2](https://github.com/vendure-ecommerce/vendure/commit/3de60b2)), closes [#2892](https://github.com/vendure-ecommerce/vendure/issues/2892)
+* **admin-ui** Generate variants even when values have not been changed ([b393538](https://github.com/vendure-ecommerce/vendure/commit/b393538)), closes [#2895](https://github.com/vendure-ecommerce/vendure/issues/2895)
+* **admin-ui** Show correct facet value language version on lang change ([7a62e7a](https://github.com/vendure-ecommerce/vendure/commit/7a62e7a)), closes [#2889](https://github.com/vendure-ecommerce/vendure/issues/2889)
+* **core** Copy prices in product duplicator (#2900) ([18d200c](https://github.com/vendure-ecommerce/vendure/commit/18d200c)), closes [#2900](https://github.com/vendure-ecommerce/vendure/issues/2900)
+* **core** Fix access to protected custom fields in Shop API ([090ff9b](https://github.com/vendure-ecommerce/vendure/commit/090ff9b)), closes [#2878](https://github.com/vendure-ecommerce/vendure/issues/2878)
+* **core** Fix memory leak cause by unbounded array growth ([5a38f42](https://github.com/vendure-ecommerce/vendure/commit/5a38f42))
+* **core** Fix regression in cookie middleware when no name provided ([fd729fd](https://github.com/vendure-ecommerce/vendure/commit/fd729fd)), closes [#2880](https://github.com/vendure-ecommerce/vendure/issues/2880)
+* **core** Support new object ref returned from plugin configuration fn ([45df738](https://github.com/vendure-ecommerce/vendure/commit/45df738)), closes [#2906](https://github.com/vendure-ecommerce/vendure/issues/2906)
+
 ## <small>2.2.5 (2024-06-03)</small>